pyproject.toml
[project]
name = "flight-tracker-led"
version = "1.0.0"
description = "Real-time overhead flight tracker for RGB LED matrix panels"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28",
    "Pillow>=9.0",
]

[project.scripts]
flight-tracker = "main:main"
